Moved fbi factory parts into fbi subfolder
